Change: Extract VersionCalculator into an own module (again)
With the removal of getting the VersionCalculator from a VersionCommand
the inter dependencies have been resolved and there isn't a cyclic
dependency anymore. Therefore VersionCalculator can live in an own
Python module again.
