+++ pontos/version/version.py
@@ -17,7 +17,6 @@
 
 from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
-from datetime import datetime
 from pathlib import Path
 from typing import Literal, Union
 
@@ -45,177 +44,6 @@ class VersionUpdate:
     previous: Version
     new: Version
     changed_files: list[Path] = field(default_factory=list)
-
-
-class VersionCalculator:
-    def next_patch_version(self, current_version: Version) -> Version:
-        """
-        Get the next patch version from a valid version
-
-        Examples:
-            "1.2.3" will return "1.2.4"
-            "1.2.3.dev1" will return "1.2.3"
-
-        Raises:
-            VersionError if version is invalid.
-        """
-
-        if current_version.is_prerelease:
-            next_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro}"
-            )
-        else:
-            next_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro + 1}"
-            )
-
-        return next_version
-
-    def next_calendar_version(self, current_version: Version) -> Version:
-        """
-        Find the correct next calendar version by checking latest version and
-        the today's date
-
-        Raises:
-            VersionError if version is invalid.
-        """
-
-        today = datetime.today()
-        current_year_short = today.year % 100
-
-        if current_version.major < current_year_short or (
-            current_version.major == current_year_short
-            and current_version.minor < today.month
-        ):
-            return parse_version(f"{current_year_short}.{today.month}.0")
-
-        if (
-            current_version.major == today.year % 100
-            and current_version.minor == today.month
-        ):
-            if current_version.dev is None:
-                release_version = parse_version(
-                    f"{current_year_short}.{today.month}."
-                    f"{current_version.micro + 1}"
-                )
-            else:
-                release_version = parse_version(
-                    f"{current_year_short}.{today.month}."
-                    f"{current_version.micro}"
-                )
-            return release_version
-        else:
-            raise VersionError(
-                f"'{current_version}' is higher than "
-                f"'{current_year_short}.{today.month}'."
-            )
-
-    def next_minor_version(self, current_version: Version) -> Version:
-        return parse_version(
-            f"{current_version.major}.{current_version.minor +1}.0"
-        )
-
-    def next_major_version(self, current_version: Version) -> Version:
-        return parse_version(f"{current_version.major + 1}.0.0")
-
-    def next_dev_version(self, current_version: Version) -> Version:
-        if current_version.is_devrelease:
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro }.dev{current_version.dev + 1}"
-            )
-        elif current_version.is_prerelease:
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro }{current_version.pre[0]}"
-                f"{current_version.pre[1]}+dev1"
-            )
-        else:
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro + 1 }.dev1"
-            )
-
-        return release_version
-
-    def next_alpha_version(self, current_version: Version) -> Version:
-        if current_version.is_devrelease:
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro }a1"
-            )
-        elif current_version.is_prerelease and current_version.pre[0] == "a":
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro }a{current_version.pre[1] + 1}"
-            )
-        else:
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro + 1 }a1"
-            )
-
-        return release_version
-
-    def next_beta_version(self, current_version: Version) -> Version:
-        if current_version.is_devrelease or (
-            current_version.is_prerelease and current_version.pre[0] == "a"
-        ):
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro }b1"
-            )
-        elif current_version.is_prerelease and current_version.pre[0] == "b":
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro }b{current_version.pre[1] + 1}"
-            )
-        else:
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro + 1 }b1"
-            )
-
-        return release_version
-
-    def next_release_candidate_version(
-        self, current_version: Version
-    ) -> Version:
-        if current_version.is_devrelease or (
-            current_version.is_prerelease and current_version.pre[0] != "rc"
-        ):
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro }rc1"
-            )
-        elif current_version.is_prerelease and current_version.pre[0] == "rc":
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro }rc{current_version.pre[1] + 1}"
-            )
-        else:
-            release_version = parse_version(
-                f"{current_version.major}."
-                f"{current_version.minor}."
-                f"{current_version.micro + 1 }rc1"
-            )
-
-        return release_version
 
 
 class VersionCommand(ABC):
